server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.16
annotated-types==0.7.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docker-pycreds==0.4.0
eval_type_backport==0.2.0
filelock==3.13.1
findspark==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.1
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
pipe==2.1
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyparsing==3.1.1
pyspark==3.5.0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.12.0
sentry-sdk==2.15.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
sympy==1.12
threadpoolctl==3.3.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchmetrics==1.4.2
torchvision==0.19.1+cu118
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wandb==0.18.3
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wkhtmltopdf==0.2
xgboost==2.0.3
yarl==1.13.1
zipp==3.17.0
